+++ app/container.py
@@ -1,17 +1,9 @@
-from fastapi import Depends, Request
 import inject
-from app.authentication import resolve_authenticated_ura_number
-from app.data import UraNumber
 from app.db.db import Database
-from app.config import Config, load_default_config
+from app.config import Config
+from app.dependencies import get_default_config
 from app.services.referral_service import ReferralService
 from app.services.pseudonym_service import PseudonymService
-
-# TODO separate the FastAPI dependencies from the injector
-
-
-def get_default_config() -> Config:
-    return load_default_config()
 
 
 def container_config(binder: inject.Binder) -> None:
@@ -36,26 +28,5 @@ def container_config(binder: inject.Binder) -> None:
     binder.bind(Config, config)
 
 
-def get_database() -> Database:
-    return inject.instance(Database)
-
-
-def get_referral_service() -> ReferralService:
-    return inject.instance(ReferralService)
-
-
-def get_pseudonym_service() -> PseudonymService:
-    return inject.instance(PseudonymService)
-
-
-def authenticated_ura(
-    request: Request, config: Config = Depends(get_default_config)
-) -> UraNumber:
-    """
-    This is a dependency being used by FastAPI.
-    """
-    return resolve_authenticated_ura_number(request, config)
-
-
-if not inject.is_configured():
-    inject.configure(container_config)
+def configure() -> None:
+    inject.configure(container_config, once=True)
