+++ app/routers/info_referrals.py
@@ -3,7 +3,7 @@ from typing import List
 from fastapi import APIRouter, Depends
 from opentelemetry import trace
 
-from app import container
+from app import dependencies
 from app.data import UraNumber
 from app.services.referral_service import ReferralService
 from app.response_models.referrals import ReferralRequest, ReferralEntry
@@ -22,9 +22,9 @@ router = APIRouter(
 )
 def get_referral_info(
     req: ReferralRequest,
-    referral_service: ReferralService = Depends(container.get_referral_service),
-    pseudonym_service: PseudonymService = Depends(container.get_pseudonym_service),
-    _: UraNumber = Depends(container.authenticated_ura),
+    referral_service: ReferralService = Depends(dependencies.get_referral_service),
+    pseudonym_service: PseudonymService = Depends(dependencies.get_pseudonym_service),
+    _: UraNumber = Depends(dependencies.authenticated_ura),
 ) -> List[ReferralEntry]:
     """
     Searches for referrals by pseudonym and data domain
